use short for U8

+++ src/main/java/dunegon/net/InputMessage.java
@@ -19,7 +19,7 @@ public class InputMessage {
         mMessageSize = messageSize;
     }
 
-    public byte getU8() {
+    public short getU8() {
         return mByteBuffer.get();
     }
